pyproject.toml
[project]
name = "genai-eval-framework"
version = "0.1.0"
description = "Evaluation framework and future gating service for Gemini- and GCP-based AI systems."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "detoxify>=0.5.0",
  "fastapi>=0.135.1",
  "google-cloud-firestore>=2.25.0",
  "google-genai>=1.68.0",
  "jinja2>=3.1.0",
  "matplotlib>=3.7.0",
  "numpy>=1.24.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.3.0",
  "scipy>=1.11.0",
  "sentence-transformers>=2.2.0",
  "torch>=2.0.0",
  "transformers>=4.35.0",
  "uvicorn>=0.42.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "src.evaluators", "src.reporters"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
